requirements.txt
# Sovereign Shadow II - Production Requirements
# Trading system with NetworkChuck strategy enforcement

# Core Framework
python>=3.11

# Data Validation & Schemas
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Exchange APIs (CCXT Library)
ccxt>=4.1.0

# Ethereum & DeFi
web3>=6.0.0
eth-account>=0.9.0
eth-utils>=2.0.0

# HTTP Requests & APIs
requests>=2.31.0
aiohttp>=3.8.0
websockets>=11.0.0

# CLI & UI
argparse  # Built-in, but listed for documentation
rich>=13.0.0  # For beautiful terminal output (optional)
tabulate>=0.9.0  # For table formatting

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Type Checking (Development)
mypy>=1.5.0
types-requests>=2.31.0
types-pyyaml>=6.0.0

# Code Quality (Development)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pylint>=2.17.0

# Logging & Monitoring
python-json-logger>=2.0.0
colorlog>=6.7.0

# Date & Time
python-dateutil>=2.8.2

# Cryptography
cryptography>=41.0.0

# JSON Processing
orjson>=3.9.0  # Fast JSON serialization (optional)

# Rate Limiting
ratelimit>=2.2.1

# API Documentation (Optional)
pydantic-openapi-schema>=1.0.0

# MCP Server (Claude Desktop Integration)
fastapi>=0.104.0  # For MCP server REST API
uvicorn>=0.24.0  # ASGI server

# Data Storage
sqlite3  # Built-in, but listed for documentation

# Utilities
attrs>=23.1.0
dataclasses-json>=0.6.0

# Obsidian Integration (Optional)
markdown>=3.4.0

# Performance
numpy>=1.24.0  # For analytics calculations
pandas>=2.0.0  # For trade journal analytics (optional)

# Environment-Specific Dependencies
# Development only - install with: pip install -r requirements-dev.txt
# pytest-xdist  # Parallel test execution
# pytest-benchmark  # Performance testing
# memory-profiler  # Memory profiling

# Production Dependencies
gunicorn>=21.0.0  # Production WSGI server (if using web interface)

# Security
certifi>=2023.0.0  # SSL certificates

# API Keys Management
keyring>=24.0.0  # Secure credential storage (optional)

# Backup & Restore
python-gnupg>=0.5.0  # GPG encryption for backups (optional)

# Notification Systems (Optional)
# discord-webhook>=1.0.0  # Discord notifications
# twilio>=8.0.0  # SMS notifications
# python-telegram-bot>=20.0  # Telegram notifications

# Chart Analysis (Future)
# matplotlib>=3.7.0  # Chart visualization
# plotly>=5.17.0  # Interactive charts
# pillow>=10.0.0  # Image processing

# Machine Learning (Future - AI trade analysis)
# scikit-learn>=1.3.0
# tensorflow>=2.14.0

# === INSTALLATION NOTES ===
#
# Basic Installation:
#   pip install -r requirements.txt
#
# Development Installation:
#   pip install -r requirements.txt
#   pip install -r requirements-dev.txt
#
# Production Installation:
#   pip install -r requirements.txt --no-dev
#
# Upgrade All:
#   pip install -r requirements.txt --upgrade
#
# Freeze Current Environment:
#   pip freeze > requirements-lock.txt
#
# === EXCHANGE-SPECIFIC NOTES ===
#
# CCXT (Cryptocurrency Exchange Trading Library)
# - Supports 100+ exchanges with unified API
# - Used for: Coinbase, Binance US, OKX, Kraken
#
# Web3.py (Ethereum Blockchain)
# - Used for: AAVE v3 position monitoring
# - Requires Ethereum RPC provider (Infura, Alchemy)
#
# Blockchain Explorers (Ledger)
# - Uses requests library for API calls
# - Etherscan API for ETH balance
# - Blockchain.info API for BTC balance
#
# === SYSTEM REQUIREMENTS ===
#
# Python: 3.11 or higher
# OS: macOS, Linux, Windows
# Memory: 2GB RAM minimum
# Storage: 1GB for logs and journal data
#
# === EXTERNAL SERVICES ===
#
# Required:
# - Exchange API keys (Coinbase, Binance US, OKX, Kraken)
# - Ethereum RPC provider (Infura/Alchemy) for AAVE
# - Etherscan API key for Ledger ETH balance
#
# Optional:
# - Discord webhook for notifications
# - Obsidian vault for configuration storage
# - Claude Desktop for MCP integration
